pyproject.toml
[tool.poetry]
name = "pgpt"
version = "0.1.0"
description = ""
authors = ["Caio Barbieri"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
openai = "^1.35.13"
python-dotenv = "^1.0.1"
flask-sqlalchemy = "^3.1.1"
requests = "^2.32.3"
pytz = "^2024.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
